+++ test/cloudevents-fast-schema-compilation.test.js
@@ -43,9 +43,16 @@ const {
   // ceArrayData
 } = require('./common-test-data')
 
+function ceValidateAlwaysFail (schema) {
+  return {
+    valid: false, // always fail
+    errors: [] // but provide no details (the same from ajv will set null here)
+  }
+}
+
 /** @test {fastifyCloudEvents} */
-test('ensure CloudEvent schema (exposed by the plugin) pass validation with a schema compiler', (t) => {
-  t.plan(29)
+test('ensure normal instancing of fast validation (like the one exposed by the plugin) is good', (t) => {
+  t.plan(30)
   const fastify = Fastify()
   t.tearDown(fastify.close.bind(fastify))
   fastify.register(require('../src/plugin')) // configure this plugin with its default options
@@ -57,22 +64,21 @@ test('ensure CloudEvent schema (exposed by the plugin) pass validation with a sc
     t.ok(CloudEvent)
     const ceSerializeFast = fastify.cloudEventSerializeFast
     t.ok(ceSerializeFast)
+    t.strictEqual(typeof ceSerializeFast, 'function')
 
     const ceSchema = fastify.cloudEventJSONSchema
     t.ok(ceSchema)
     t.strictEqual(typeof ceSchema, 'object')
 
+    // instancing schema compiler in the same way of the plugin,
+    // but in a manual way here, as a sample
     const ajv = new Ajv({ coerceTypes: true, removeAdditional: true })
     assert(ajv !== null)
     t.ok(ajv)
     const ceValidate = ajv.compile(ceSchema)
     assert(ceValidate !== null)
-    assert(typeof ceValidate === 'function')
     t.ok(ceValidate)
-    const ceValidateAlwaysFail = function (schema) {
-      return false // always fail
-    }
-    t.ok(ceValidateAlwaysFail)
+    t.strictEqual(typeof ceValidate, 'function')
 
     // test on some good data
     const ceFullStrict = new CloudEvent('1/full/sample-data/strict',
@@ -125,16 +131,16 @@ test('ensure CloudEvent schema (exposed by the plugin) pass validation with a sc
       // serialization and validation tests on the good test object
       const ceFullStrictSerializedFast = ceSerializeFast(ceFullStrict, { onlyValid: true })
       t.ok(ceFullStrictSerializedFast)
-      const ceFullStrictValid = ceValidateAlwaysFail(ceFullStrict)
-      // if (!ceFullStrictValid) console.log(`DEBUG: validation errors: ${JSON.stringify(ceValidateAlwaysFail.errors)}`)
-      t.ok(!ceFullStrictValid)
+      const ceFullStrictValidated = ceValidateAlwaysFail(ceFullStrict)
+      // console.log(`DEBUG - ceFullStrict, validation by always fail validator: ${JSON.stringify(ceFullStrictValidated)}`)
+      t.ok(!ceFullStrictValidated.valid)
 
       // serialization and validation tests on the bad test object
       const ceFullBadSerializedOnlyValidFalse = ceSerializeFast(ceFullBad, { onlyValid: false })
       t.ok(ceFullBadSerializedOnlyValidFalse)
-      const ceFullBadValid = ceValidateAlwaysFail(ceFullBad)
-      // if (!ceFullBadValid) console.log(`DEBUG: validation errors: ${JSON.stringify(ceValidateAlwaysFail.errors)}`)
-      t.ok(!ceFullBadValid)
+      const ceFullBadValidated = ceValidateAlwaysFail(ceFullBad)
+      // console.log(`DEBUG - ceFullBad, validation by always fail validator: ${JSON.stringify(ceFullBadValidated)}`)
+      t.ok(!ceFullBadValidated.valid)
     }
 
     {
@@ -161,6 +167,104 @@ test('ensure CloudEvent schema (exposed by the plugin) pass validation with a sc
       //   console.log(`DEBUG - ceStrict, dump validation errors with ajv: ${JSON.stringify(ceValidate.errors)}`)
       // }
       t.ok(ceStrictValid)
+      // console.log(`DEBUG - ceStrict serialized:\n${ceStrictSerializedFast}`)
+      const ceStrictAsString = ceStrict.toString()
+      // console.log(`DEBUG - ceStrictAsString: ${ceStrictAsString}`)
+      t.ok(ceStrictAsString != null && (typeof ceStrictAsString === 'string'))
+      const ceStrictPayloadDumped = JSON.stringify(ceStrict.payload)
+      // console.log(`DEBUG - ceStrictPayloadDumped: ${ceStrictPayloadDumped}`)
+      t.ok(ceStrictPayloadDumped != null && (typeof ceStrictPayloadDumped === 'string'))
+      t.ok(ceStrictPayloadDumped.length < 1024)
+    }
+  })
+})
+
+/** @test {fastifyCloudEvents} */
+test('ensure CloudEvent schema and schema compiler (both exposed by the plugin) pass validation', (t) => {
+  t.plan(26)
+  const fastify = Fastify()
+  t.tearDown(fastify.close.bind(fastify))
+  fastify.register(require('../src/plugin')) // configure this plugin with its default options
+
+  fastify.listen(0, (err, address) => {
+    t.error(err)
+
+    const CloudEvent = fastify.CloudEvent
+    t.ok(CloudEvent)
+    const ceSerializeFast = fastify.cloudEventSerializeFast
+    t.ok(ceSerializeFast)
+    t.strictEqual(typeof ceSerializeFast, 'function')
+
+    const ceSchema = fastify.cloudEventJSONSchema
+    t.ok(ceSchema)
+    t.strictEqual(typeof ceSchema, 'object')
+
+    const ceValidateFast = fastify.cloudEventValidateFast
+    t.ok(ceValidateFast)
+    t.strictEqual(typeof ceValidateFast, 'function')
+
+    // test on some good data
+    const ceFullStrict = new CloudEvent('1/full/sample-data/strict',
+      ceNamespace,
+      ceServerUrl,
+      ceCommonData, // data
+      ceCommonOptionsStrict,
+      ceCommonExtensions
+    )
+    t.ok(ceFullStrict)
+    t.ok(ceFullStrict.isValid())
+
+    // additional tests, with bad objects ...
+    const ceFullBad = new CloudEvent(null,
+      ceNamespace,
+      ceServerUrl,
+      ceCommonData, // data
+      ceCommonOptions,
+      {} // extensions
+    )
+    t.ok(ceFullBad)
+    t.ok(!ceFullBad.isValid())
+
+    {
+      // tests using the good validator
+      // serialization and validation tests on the good test object
+      // console.log(`DEBUG - dump standard validation errors: ${CloudEvent.dumpValidationResults(ceFullStrict, {}, 'ceFullStrict')}`)
+      const ceFullStrictSerializedFast = ceSerializeFast(ceFullStrict, { onlyValid: true })
+      t.ok(ceFullStrictSerializedFast)
+      const ceFullStrictValidated = ceValidateFast(ceFullStrict)
+      // console.log(`DEBUG - ceFullStrict, validation by ajv: ${JSON.stringify(ceFullStrictValidated)}`)
+      t.ok(ceFullStrictValidated.valid)
+
+      // serialization and validation tests on the bad test object
+      // console.log(`DEBUG - dump standard validation errors: ${CloudEvent.dumpValidationResults(ceFullBad, {}, 'ceFullBad')}`)
+      const ceFullBadSerializedOnlyValidFalse = ceSerializeFast(ceFullBad, { onlyValid: false })
+      t.ok(ceFullBadSerializedOnlyValidFalse)
+      const ceFullBadValidated = ceValidateFast(ceFullBad)
+      // console.log(`DEBUG - ceFullBad, validation by ajv: ${JSON.stringify(ceFullBadValidated)}`)
+      t.ok(!ceFullBadValidated.valid)
+    }
+
+    {
+      const value = 'Hello World, 2020'
+      // use directly the event with strict mode enabled ...
+      const ceStrict = new CloudEvent('1/full/string-data-text-mime-type/strict',
+        ceNamespace,
+        ceServerUrl,
+        value, // data
+        ceCommonOptionsForTextDataStrict,
+        ceCommonExtensions
+      )
+      assert(ceStrict !== null)
+      t.ok(ceStrict)
+      t.ok(CloudEvent.isValidEvent(ceStrict))
+      t.strictSame(ceStrict.payload, ceStrict.data)
+      t.strictSame(ceStrict.dataType, 'Text')
+      const ceStrictSerializedFast = ceSerializeFast(ceStrict, { onlyValid: true })
+      t.ok(ceStrictSerializedFast)
+      const ceStrictValidated = ceValidateFast(ceStrict)
+      // console.log(`DEBUG - ceStrict, validation by ajv: ${JSON.stringify(ceStrictValidated)}`)
+      t.ok(ceStrictValidated)
+      t.ok(ceStrictValidated.valid)
       // console.log(`DEBUG - ceStrict serialized:\n${ceStrictSerializedFast}`)
       const ceStrictAsString = ceStrict.toString()
       // console.log(`DEBUG - ceStrictAsString: ${ceStrictAsString}`)