+++ CHANGELOG.md
@@ -13,6 +13,8 @@ Summary Changelog:
   from the underlying library
 - Feature: update fast serialization to handle even data as value (string or boolean or number) 
   with datacontenttype not default (for example 'text/plain')
+- Feature: add to Fastify a decorator to add a function that validates with a schema compiler 
+  and return validation results and errors (if any) 
 
 ## [2.5.0](https://github.com/smartiniOnGitHub/fastify-cloudevents/releases/tag/2.5.0) (2020-10-02)
 [Full Changelog](https://github.com/smartiniOnGitHub/fastify-cloudevents/compare/2.4.0...2.5.0)